Refactor Neato botvac components as a vacuum
A switch is still use to enable/disable the schedule
