+++ homeassistant/components/neato.py
@@ -90,7 +90,7 @@ def setup(hass, config):
         _LOGGER.debug("Failed to login to Neato API")
         return False
     hub.update_robots()
-    for component in ('camera', 'sensor', 'switch'):
+    for component in ('camera', 'vacuum', 'switch'):
         discovery.load_platform(hass, component, DOMAIN, {}, config)
 
     return True
